CI(azure-pipelines): use correct path for build output
"Agent.BuildDirectory" refers to the path on the agent where all folders for the pipeline are created.

"Build.BinariesDirectory" is the correct variable to use: it refers to the path that can be used as an output folder for compiled binaries.

Reference: https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables