CI(azure-pipelines): use our own macOS builder for releases
MacStadium kindly provided us an hosted 2012 Mac mini for free last year.

Until now we only used it for testing and to build the vcpkg environment, waiting for the CMake project to be merged.

Now that everything is in place, we can use it for CI and to release builds.