FEAT(talking-ui): Make channels selectable
Before this commit it was only possible to select users in the TalkingUI
(and thereby it was also only possible to trigger the context menu for
users via the TalkingUI).

This commit extends the functionality that currently exists for users to
channels as well. This includes selection as well as the context menu.