CI(azure): Use new Go AppImage tool
The old AppImage tool required to be run on the oldest still supported
Ubuntu LTS. On anything else it would quit with an error.

There is a new tool from the same author though that does not have this
limitation due to bundling more of its dependencies into the AppImage.
The tool is taken from https://github.com/probonopd/go-appimage.