FIX(client): Racecondition deleting shortcuts
Server-specific shortcuts are read into memory in
MainWindow::msgServerSync and are saved into the DB again in
MainWindow::serverDisconnected.
If however the disconnect function runs before the server
synchronization has finished, the shortcut list in memory is empty and
by writing that into the DB all existing server-specific shortcuts are
lost (overwritten).

This commit adds checks before saving the shortcuts so that saving only
occurs if the synchronization has happened before and thus the
shortcuts have been loaded in the first place.

(Backported from #4430 and adapted to work with the 1.3.x series)