FIX(client,code): UserModel::getChannel
The function first checks whether there is an user that is associated
with the given ModelIndex. If this is the case, it'll return the channel
that user is in. This however is not necessarily the channel that the
ModelIndex is grouped to.
Therefore this commit changed the function to prefer returning the
channel entry of the model's parent.