FEAT(talking-ui): Save last position
In order for the UI to not pop-up in the middle of the MainWindow (or
any other default location) after you have placed it carefully the last
time, this commit makes the UI's position to be saved and restored.

For cases in which there is no position set yet, it will be positioned
to the right of the initial position of the MainWindow.

Note that the UI's position is only saved if it is visible at the time
Mumble is closed which is used as an indication that the user is
actually using it (and has thus placed it explicitly).