FEAT(client, mainwindow): Show volume adjustments
This feature optionally allows users to display the set volume
adjustments next to the user's/listener's name.
It will only be shown if it is not equal to 0 dB (aka: it is actually
having an effect).

By default this option is enabled since it is only visible if volume
adjustments are set up anyways.