BUILD(cmake): Start cmake output with info header
Instead of only printing the Mumble version that is being built, the
architecture that cmake is currently configuring for is displayed as
well.
In order to make this information more visible, they are surrounded by a
row of hashtags.

This change should make it more obvious to users if they happen to build
the wrong version/architecture (as can easily happen on Windows).