FIX(client, ui): ACLEditor not growing evenly
The ACLEditor's Groups tab has 3 separate columns for dealing with
members. These columns wouldn't grow equally though: The 2 leftmost
columns basically didn't grow at all while the rightmost grew to fill
the available space.

This is a problem though since the 2 leftmost columns have a ComboBox in
them and if the user wants to make those bigger, then (s)he expects this
to be possible by increasing the window's size. This was not the case
though and the ComboBoxes stayed at their initial size.

This commit makes sure that all columns now grow equally and thus
allowing the user to increase the ComboBoxe's size by increasing the
window's size.

Fixes #4453