MAINT: Use rel paths in updatetranslations.sh
The script now uses relative paths (in order to not include information
about the committer's file tree in the commit message.

Furthermore the pushd/popd was removed as we don't have to restore the
path in a shell script as these run in a sub-shell anyways and therefore
can't change the calling shell's pwd.