BUILD(cmake): Fix opus library not found
When trying to build Mumble using a non-bundled Opus version, the build
would fail as cmake fails to find the respective library. This is due to
the library name in the cmake file starting with a capital letter
instead of a lowercase one (as it does in the actual library's name).