BUILD(cmake): Select tests based on built components
As some tests only make sense to be run if a given component is included
in the build (client/server), this commit makes sure no unnecessary
tests are built (and run).

Furthermore it removed the dependency of some tests on a certain target
having a source-dir property.

And finally it also removed the references to "NO_USE_SRV" from
TestServerResolver as this macro is never defined anywhere (it's
probably a remnant of the old qmake system).

Co-Authored-By: Sebastian Parborg <darkdefende@gmail.com>