FIX(client, audio): Adapt search-path: Opus + CELT
The default install directories for the Opus and CELT directories have
been changed for Unix systems. Instead of installing in the normal
library-location they will now be installed in a mumble/ subdirectory in
order to not overwrite existing libraries.

In order for Mumble to actually find them, the search-algorithm used to
locate the libraries had to be adjusted as well, which is exactly what
this commit does.