FIX(client,ui): MainWindow not updating on vol. adj.
If the local volume adjustment of a client or a ChannelListener was
changed, the MainWindow would not reflect this change until some event
triggered it to refresh (e.g. MouseOver).

With this fix, the MainWindow gets updated immediately when the volume
adjustment is changed (of a client or a ChannelListener).