FIX(client, ui): Make setting take effect immediately
The setting for keeping the local user always visible in the TalkingUI
was not put into effect immediately if the local user didn't happen to
be displayed in the TalkingUI at the time of applying the settings.
Instead the effect was only visible after a re-connect or after having
changed the local user's TalkingState so that it was added to the UI via
that route.

The fix was simply to add the user if it is currently not present but is
asked to be shown permanently.