REFAC(overlay): improve find_odlsym(), fix crash when DT_HASH is not present
dlsym() is changed so that it returns NULL when the original dlsym() (odlsym()) is not available.

find_odlsym() is refactored:

- A segmentation fault was reported many years ago on Gentoo's bug tracker: https://bugs.gentoo.org/527504
Turns out it happens when DT_HASH is not available. The function didn't check whether the required tables existed.
This commit fixes the crash. Please note that support for DT_GNU_HASH is not added, it will be in a future commit.

- The function currently supports Linux and FreeBSD. Only the way the library is opened and some types change.
Because of that, this commit attempts to reduce the amount of duplicate code as much as possible.

- Correct types are now used, e.g. "uint32_t" instead of "int" for the number of chains.

- Other minor changes.