FIX(talking-ui): Context-menu z-index issues
In cases where the TalkingUI was overlaying the MainWindow, it could
happen that a context menu triggered from the TalkingUI would actually
be created using another element (listener, channel, user) as the
context for that menu. This would happen because in these cases there is
an entry below the cursor (behind the TalkingUI) that would
automatically be recognized as the desired target.

The fix works by explicitly disabling the position-dependent target
search and using explicitly defined targets instead.