FEAT(client, ui): Add button to reset all settings
Previously the user only had the choice of setting the settings on the
current settings page to their default value. Therefore, if you wanted
to reset all settings, you had to visit every page and reset it
individually.

With this commit, there is now a reset-all button that can be used
instead.

Co-authored-by: Davide Beatrici <github@davidebeatrici.dev>