FIX(client): don't print Zeroconf warning if the operation was cancelled
Closing ConnectDialog causes the ZeroConf object to be destroyed. The destructor cancels any running operations, as expected.

However, a warning is printed in Zeroconf::callbackBrowseComplete() because ERROR_CANCELLED is not handled:

"Zeroconf: DnsServiceBrowse() reports status code 1223, ignoring results"

This commit fixes the issue.

419b06065fe7bf1b3133aff9b29002fce71806b8 applied the exact same change to callbackResolveComplete().

The error message in that commit's description is wrong, but everything else is correct.