DOCS: Improve CMake install instructions
We require CMake in at least version 3.5 and mention this.

All supported Ubuntu versions (starting with Xenial) provide the cmake package in version 3.5.1 or higher. [1]

All supported Debian versions (starting with 'oldstable') provide the cmake package in version 3.7.2 or higher. [2]

The upstream CMake PPA[3] only provides PPAs for Ubuntu Xenial and up.

Hence we can depend on cmake package versions usually being 3.5 or up for Linux distributions. Listing the PPA does not help for older Ubuntu versions where using it would be necessary. And it does not make sense for non-Ubuntu distributions as only Ubuntu PPAs are provided.

[1] https://packages.ubuntu.com/search?keywords=cmake&searchon=names&exact=1&suite=all&section=main
[2] https://packages.debian.org/search?keywords=cmake&searchon=names&exact=1&suite=oldstable&section=all
[3] https://apt.kitware.com/