FIX(positional-audio): Wrong character size used for wide strings
"wchar_t" is usually 4 bytes big. That's not the case on Windows, where it's 2 bytes instead.

We cannot rely on that type on Linux because the target process could be native or running through Wine.

Also, it's possible that a game uses the same size for "wchar_t" on both Linux and Windows.

This commit introduces two variants of utf16ToUtf8():

1. Accepts std::u16string, for Windows processes.
2. Accepts std::u32string, for Linux processes.

The old variant, which accepted std::wstring, is removed to prevent misuses.