FIX(client, talking-ui): Icon size initialized incorrectly
If the font size for the TalkingUI was set to something different than
100%, the TalkingUI's icon sizes would be incorrect.
This is due to the icon size being initialized to the widget's
font-size. This doesn't seem to have been affected by inherited
StyleSheets right after construction and therefore they'd always report
the default font-size.

This commit fixes this by explicitly setting the icon-size after an
entry is created.

In this context the method of calculating the current line height has
been overhauled as well as this would also depend on when the respective
function was called. This is probably due to Qt not having stored the
display's DPI correctly yet and thus the conversion from point size to
pixel size is flawed.