FEAT(client): Add possibility to set nickname
As requested in #3967 this commit implements the ability to set
nicknames for other users when connected to the same server. The change
is added to the client database by the user id for persistence. In order
to set a nickname, a user can right click on that user and click Set
Nickname. This opens a window where the user can enter the new nickname.
If the text field is cleared, no nickname is displayed.

Implements #3967