CI(azure): Get unique build number from our server
BUILD_NUMBER_TOKEN is defined on Azure Pipelines and contains a secret token that is used to access the endpoint on our server.

Restricted access is in place for the sole purpose of preventing the build number from being increased when accessing the page outside of the CI environment.

This commit also improves the descriptive header of the build script files.