FEAT(client, mac): Prompt user for microphone permission if not granted
This commit utilized the internal permission check function to examine
if the user has granted mumble the microphone permission when initializing
CoreAudioInput backend. If the user hasn't granted the permission, it will
make a request through macOS's interface. If being denied, a warning message
will appear in the chat log.

This commit addressed #4547 and #4580. However, self-compiled unsigned
mumble binary still cannot access microphone because this is denied by
Apple's policy. The workaround is to start mumble binary from a terminal,
which is properly signed (e.g. the Terminal app pre-installed).

Implement #4547 and #4580.