FIX(client,ui): Channel description via lock icon
With the introduction of #3929 a channel's description would not be
accessible by clicking on the comment icon next to the channel's name
but rather by clicking to the lock icon to its right (if there was one).
This was due to some implicit assumptions about the comment icon's
position in the code that is responsible for showing the description.

Thus the fix was simply to let that part of the code account for the
possible new icon as well.

Fixes #4685

Co-Authored-By: Davide Beatrici <git@davidebeatrici.dev>